[Docs] Move out the react part of the index
