+++ docs/src/app.native.js
@@ -1,0 +1,31 @@
+import React, {
+  AppRegistry,
+  StyleSheet,
+  ScrollView,
+  StatusBar,
+} from 'react-native';
+
+import Main from './Main.native';
+
+const styles = StyleSheet.create({
+  root: {
+    backgroundColor: '#fff',
+  },
+});
+
+const App = () => (
+  <ScrollView style={styles.root}>
+    <StatusBar
+      backgroundColor="#0F3D6C"
+      translucent={true}
+      barStyle="light-content"
+    />
+    <Main />
+  </ScrollView>
+);
+
+export default {
+  start() {
+    AppRegistry.registerComponent('SwipeableViewsDocs', () => App);
+  },
+};
