+++ index.ios.js
@@ -1,7 +1,3 @@
-import {
-  AppRegistry,
-} from 'react-native';
+import app from './docs/src/app.native';
 
-import Main from './docs/src/Main.native';
-
-AppRegistry.registerComponent('SwipeableViewsDocs', () => Main);
+app.start();
